# cmd/swag/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.LoginInput:
    properties:
      user_account:
        type: string
      user_password:
        type: string
    type: object
  model.RegisterInput:
    properties:
      verify_key:
        type: string
      verify_purpose:
        type: string
      verify_type:
        type: integer
    type: object
  model.SetupTwoFactorAuthInput:
    properties:
      two_factor_auth_type:
        type: string
      two_factor_email:
        type: string
      user_id:
        type: integer
    type: object
  model.TwoFactorVerificationInput:
    properties:
      two_factor_code:
        type: string
      user_id:
        type: integer
    type: object
  model.UpdatePasswordRegisterInput:
    properties:
      user_password:
        type: string
      user_token:
        type: string
    type: object
  model.VerifyInput:
    properties:
      verify_code:
        type: string
      verify_key:
        type: string
    type: object
  response.ErrorResponseData:
    properties:
      code:
        description: status code
        type: integer
      detail:
        description: du lai return
      error:
        description: thong bao loi
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        description: status code
        type: integer
      data:
        description: dữ liệu return
      message:
        description: thông báo lỗi
        type: string
    type: object
host: localhost:8002
info:
  contact:
    email: phuocsanhtps@gmail.com
    name: API Support
    url: https://github.com/phuocsanh/Go_Ecommerce-go
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/phuocsanh/Go_Ecommerce-go
  title: Api documentation ecommerce_sq
  version: "1.0"
paths:
  /user/login:
    post:
      consumes:
      - application/json
      description: User Login
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: User Login
      tags:
      - account management
  /user/register:
    post:
      consumes:
      - application/json
      description: When user is registered send otp to email
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: User Registration
      tags:
      - account management
  /user/two-factor/setup:
    post:
      consumes:
      - application/json
      description: ser Setup Two Factor Authentication
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.SetupTwoFactorAuthInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: ser Setup Two Factor Authentication
      tags:
      - account 2fa
  /user/two-factor/verify:
    post:
      consumes:
      - application/json
      description: ser Verify Two Factor Authentication
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.TwoFactorVerificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: ser Verify Two Factor Authentication
      tags:
      - account 2fa
  /user/update_pass_register:
    post:
      consumes:
      - application/json
      description: UpdatePasswordRegister
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePasswordRegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: UpdatePasswordRegister
      tags:
      - account management
  /user/verify_account:
    post:
      consumes:
      - application/json
      description: Verify OTP Login By User
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.VerifyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Verify OTP Login By User
      tags:
      - account management
swagger: "2.0"